% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarmaRidgeBJPS2025-package.R
\docType{package}
\name{BarmaRidgeBJPS2025-package}
\alias{BarmaRidgeBJPS2025}
\alias{BarmaRidgeBJPS2025-package}
\title{BarmaRidgeBJPS2025: Enhancements for Numerical Stability in Beta ARMA Models}
\description{
Implements a ridge penalization scheme to improve the numerical stability of conditional maximum likelihood estimation for Beta Autoregressive Moving Average (BARMA) models. The penalty term enhances the curvature of the log-likelihood function, reducing the risk of convergence failures and implausible estimates. The package also provides a bootstrap-based parameter estimation strategy as a complementary solution. These tools are designed to replicate the analysis and methods presented in Cribari-Neto, Costa, and Fonseca (2025) \doi{10.1214/25-BJPS645}.
}
\references{
Cribari-Neto, F., Costa, E., & Fonseca, R. V. (2025). 
Numerical stability enhancements in beta autoregressive moving average model
 estimation. \emph{Brazilian Journal of Probability and Statistics}, 39(4), 
 410-437. <doi:10.1214/25-BJPS645>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Everton-da-Costa/BarmaRidgeBJPS2025}
  \item Report bugs at \url{https://github.com/Everton-da-Costa/BarmaRidgeBJPS2025/issues}
}

}
\author{
\strong{Maintainer}: Everton Costa \email{everto.cost@gmail.com} (\href{https://orcid.org/0000-0001-7580-2639}{ORCID})

Other contributors:
\itemize{
  \item Rodney V. Fonseca \email{rodneyfonseca@ufba.br} (\href{https://orcid.org/0000-0003-3948-3145}{ORCID}) [contributor]
  \item Francisco Cribari-Neto \email{francisco.cribari@ufpe.br} (\href{https://orcid.org/0000-0002-5909-6698}{ORCID}) [contributor]
}

}
\keyword{internal}
