% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_seasonal.R
\name{bootstrap_seasonal}
\alias{bootstrap_seasonal}
\title{Block Bootstrap for Seasonal Time Series (Dudek et al., 2013)}
\usage{
bootstrap_seasonal(y, block_length, seasonal_period = 12)
}
\arguments{
\item{y}{A time series object containing the original time series data.}

\item{block_length}{An integer specifying the block size (`b`),
which determines the length of contiguous observations sampled together
during resampling.}

\item{seasonal_period}{An integer specifying the seasonal period length
(`d`), such as 12 (for monthly data with yearly seasonality) or
24, 36, etc.}
}
\value{
A list with two elements:
\describe{
  \item{y_star_ts}{A time series object (`ts`) containing the bootstrapped
  time series data.}
  \item{ids_y_bootstrap}{A numeric vector containing the indices of the
  sampled blocks for each bootstrap iteration.}
}
}
\description{
Implements the generalized block bootstrap method for seasonal time series
data, as described in Dudek et al. (2013). This method is designed to
resample seasonal time series in a way that preserves its periodic structure.
}
\details{
This function implements the following steps (from Dudek et al., 2013):

- **Step 1**: Define block size (`b`) and calculate the number of blocks
   (`l`). Identify the set of starting indices for the blocks.

- **Step 2**: For each block starting index (`t`), calculate the set of
  possible indices (`S_\{t,n\}`) from which blocks can be sampled.
  This ensures that the seasonal structure is preserved.

- **Step 3**: Randomly sample blocks from the time series based on the sets
  calculated in Step 2, and assemble the bootstrapped series by joining
  these blocks.

The method ensures that the seasonal dependencies in the data are maintained
by carefully defining the set of candidate indices for resampling.
}
\references{
Dudek, A. E., Leskow, J., Paparoditis, E., & Politis, D. N. (2014).
A GENERALIZED BLOCK BOOTSTRAP FOR SEASONAL TIME SERIES.
_Journal of Time Series Analysis_, 35(2), 89–114. DOI: 10.1111/jtsa.12053
}
\keyword{internal}
