% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brasilia_ts}
\alias{brasilia_ts}
\title{Brasília Relative Humidity Monthly Time Series}
\format{
A time series object (`ts`) with 306 monthly observations from
January 1999 to June 2024. [cite: 37]
}
\source{
The data were obtained from the NASA (National Aeronautics and Space
Administration) Prediction of Worldwide Energy Resources (POWER) project. [cite: 38]
}
\usage{
brasilia_ts
}
\description{
Monthly relative humidity data from Brasília, Brazil, as a time series object,
starting from January 1999. [cite: 37]
}
\keyword{datasets}
