% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_model.R
\name{inf_matrix_arma}
\alias{inf_matrix_arma}
\title{Fisher Information Matrix for a BARMA Model}
\usage{
inf_matrix_arma(y, ar, ma, alpha = 0, varphi = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in the open interval (0,1).}

\item{ar}{A numeric vector specifying the autoregressive (AR) lags.}

\item{ma}{A numeric vector specifying the moving average (MA) lags.}

\item{alpha}{The intercept term.}

\item{varphi}{A numeric vector of autoregressive (AR) parameters.}

\item{theta}{A numeric vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the BARMA model.}

\item{link}{The link function to be used.}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted': The fitted values (conditional mean).
  - 'muhat': The truncated vector of fitted conditional means.
  - 'etahat': The full vector of the fitted linear predictor.
  - 'errorhat': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix of the
  Beta Autoregressive Moving Average (BARMA) model. It also computes
  auxiliary values like fitted values and residuals.
}
\keyword{internal}
