% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax_model.R
\name{inf_matrix_armax}
\alias{inf_matrix_armax}
\title{Fisher Information Matrix for a BARMAX Model}
\usage{
inf_matrix_armax(
  y,
  ar,
  ma,
  X,
  alpha = 0,
  varphi = 0,
  theta = 0,
  phi = 0,
  beta = 0,
  link = "logit"
)
}
\arguments{
\item{y}{Data, a time series of numbers in the open interval (0,1).}

\item{ar}{A numeric vector specifying the autoregressive (AR) lags.}

\item{ma}{A numeric vector specifying the moving average (MA) lags.}

\item{X}{A matrix of exogenous variables (regressors).}

\item{alpha}{The intercept term of the linear predictor.}

\item{varphi}{A numeric vector of autoregressive (AR) parameters.}

\item{theta}{A numeric vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A numeric vector of regression coefficients for X.}

\item{link}{The link function to be used.}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted': The fitted values (conditional mean).
  - 'muhat': The truncated vector of fitted conditional means.
  - 'etahat': The full vector of the fitted linear predictor.
  - 'errorhat': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix for the
  BARMAX model.
}
\keyword{internal}
