% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax_model.R
\name{inf_matrix_armax_ridge}
\alias{inf_matrix_armax_ridge}
\title{Penalized Fisher Information Matrix for a BARMAX Model}
\usage{
inf_matrix_armax_ridge(
  y,
  ar,
  ma,
  X,
  link,
  alpha,
  varphi,
  theta,
  phi,
  beta,
  penalty
)
}
\arguments{
\item{y}{A time series of numbers in the interval (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) lags.}

\item{ma}{A vector specifying the moving average (MA) lags.}

\item{X}{A matrix of external regressors.}

\item{link}{The link function ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter (not penalized).}

\item{beta}{A vector of regression coefficients (not penalized).}

\item{penalty}{The non-negative L2 penalty parameter (lambda).}
}
\value{
A list containing the penalized FIM and other model outputs.
}
\description{
This function computes the L2-penalized FIM for a BARMAX model.
}
\keyword{internal}
