% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_model.R
\name{inf_matrix_arx}
\alias{inf_matrix_arx}
\title{Fisher Information Matrix for a BARX Model}
\usage{
inf_matrix_arx(y, ar, X, alpha = 0, varphi = 0, phi = 0, beta = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in the open interval (0,1).}

\item{ar}{A numeric vector specifying the autoregressive (AR) lags.}

\item{X}{A matrix of external regressors.}

\item{alpha}{The intercept term of the linear predictor.}

\item{varphi}{A numeric vector of autoregressive (AR) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A numeric vector of parameters for the external regressors.}

\item{link}{The link function to be used.}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted': The fitted values (conditional mean).
  - 'muhat': The truncated vector of fitted conditional means.
  - 'etahat': The full vector of the fitted linear predictor.
  - 'errorhat': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix of the
  Beta Autoregressive model with exogenous regressors (BetaARX).
}
\keyword{internal}
