% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_model.R
\name{inf_matrix_arx_ridge}
\alias{inf_matrix_arx_ridge}
\title{Penalized Fisher Information Matrix for a BARX Model}
\usage{
inf_matrix_arx_ridge(y, ar, X, link, alpha, varphi, phi, beta, penalty)
}
\arguments{
\item{y}{A time series of data in the (0,1) interval.}

\item{ar}{A vector specifying the autoregressive (AR) lags.}

\item{X}{A matrix of external regressors.}

\item{link}{The link function ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{phi}{The precision parameter (not penalized).}

\item{beta}{A vector of regression coefficients (not penalized).}

\item{penalty}{The non-negative L2 penalty parameter (lambda).}
}
\value{
A list containing the penalized FIM and other model outputs.
}
\description{
This function computes the L2-penalized Fisher Information
  Matrix (FIM) for a Beta Autoregressive model with external regressors.
}
\keyword{internal}
