% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_model.R
\name{inf_matrix_ma}
\alias{inf_matrix_ma}
\title{Fisher Information Matrix for a BMA Model}
\usage{
inf_matrix_ma(y, ma, alpha = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in the open interval (0,1).}

\item{ma}{A numeric vector specifying the moving average (MA) lags.}

\item{alpha}{The intercept term.}

\item{theta}{A numeric vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the BMA model.}

\item{link}{The link function to be used.}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted': The fitted values (conditional mean).
  - 'muhat': The truncated vector of fitted conditional means.
  - 'etahat': The full vector of the fitted linear predictor.
  - 'errorhat': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix of the
  Beta Moving Average (BMA) model.
}
\keyword{internal}
