% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_model.R
\name{inf_matrix_ma_ridge}
\alias{inf_matrix_ma_ridge}
\title{Penalized Fisher Information Matrix for a BMA Model}
\usage{
inf_matrix_ma_ridge(y, ma, link, alpha, theta, phi, penalty)
}
\arguments{
\item{y}{A time series of data in the (0,1) interval.}

\item{ma}{A vector specifying the moving average (MA) lags.}

\item{link}{The link function ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter (not penalized).}

\item{penalty}{The non-negative L2 penalty parameter (lambda).}
}
\value{
A list containing the penalized FIM and other model outputs.
}
\description{
This function computes the L2-penalized Fisher Information
  Matrix (FIM) for a Beta Moving Average (BMA) model.
}
\keyword{internal}
