% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_model.R
\name{inf_matrix_max}
\alias{inf_matrix_max}
\title{Fisher Information Matrix for a BMAX Model}
\usage{
inf_matrix_max(y, ma, X, alpha = 0, theta = 0, phi = 0, beta = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in the open interval (0,1).}

\item{ma}{A numeric vector specifying the moving average (MA) lags.}

\item{X}{A matrix of exogenous variables (regressors).}

\item{alpha}{The intercept term of the linear predictor.}

\item{theta}{A numeric vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A numeric vector of regression coefficients.}

\item{link}{The link function to be used ("logit", "probit", etc.).}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted': A time series object of the fitted conditional mean values.
  - 'muhat': The truncated vector of fitted conditional mean values.
  - 'etahat': The full vector of the linear predictor values.
  - 'errorhat': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix for the
  Beta Moving Average model with exogenous regressors (BMAX).
}
\keyword{internal}
