% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_model.R
\name{inf_matrix_max_ridge}
\alias{inf_matrix_max_ridge}
\title{Penalized Fisher Information Matrix for a BMAX Model}
\usage{
inf_matrix_max_ridge(y, ma, X, link, alpha, theta, phi, beta, penalty)
}
\arguments{
\item{y}{A time series of data in the (0,1) interval.}

\item{ma}{A vector specifying the moving average (MA) lags.}

\item{X}{A matrix of external regressors.}

\item{link}{The link function ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter (not penalized).}

\item{beta}{A vector of regression coefficients (not penalized).}

\item{penalty}{The non-negative L2 penalty parameter (lambda).}
}
\value{
A list containing the penalized FIM and other model outputs.
}
\description{
This function computes the L2-penalized Fisher Information
  Matrix (FIM) for a BMAX model. The penalty is applied to the intercept
  (alpha) and the MA parameters (theta).
}
\keyword{internal}
