% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_model.R
\name{loglik_ar}
\alias{loglik_ar}
\title{Log-Likelihood for a BAR Model}
\usage{
loglik_ar(y, ar, alpha = 0, varphi = 0, phi = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) component.}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{link}{The link function to be used ("logit", "probit", "loglog",
"cloglog").}
}
\value{
A numeric value representing the total conditional log-likelihood
  of the BAR model for the given data and parameters.
}
\description{
This function computes the log-likelihood of the Beta
  Autoregressive (BAR) model.
}
\keyword{internal}
