% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_model.R
\name{loglik_arma}
\alias{loglik_arma}
\title{Log-Likelihood for a BARMA Model}
\usage{
loglik_arma(y, ar, ma, alpha = 0, varphi = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) component.}

\item{ma}{A vector specifying the moving average (MA) component.}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the BARMA model.}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}
}
\value{
The likelihood of the BARMA estimators.
}
\description{
This function computes the log-likelihood of the BARMA model.
}
\keyword{internal}
