% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax_model.R
\name{loglik_armax_ridge}
\alias{loglik_armax_ridge}
\title{Penalized Log-Likelihood for a BARMAX Model}
\usage{
loglik_armax_ridge(
  y,
  ar,
  ma,
  X,
  link,
  alpha,
  varphi,
  theta,
  phi,
  beta,
  penalty
)
}
\arguments{
\item{y}{A time series of numbers in the interval (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) lags.}

\item{ma}{A vector specifying the moving average (MA) lags.}

\item{X}{A matrix of external regressors.}

\item{link}{The link function to be used ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term of the linear predictor.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the beta distribution.}

\item{beta}{A vector of regression coefficients for external regressors.}

\item{penalty}{A non-negative scalar for the L2 penalty.}
}
\value{
The value of the L2 penalized log-likelihood.
}
\description{
Computes the L2-penalized (Ridge) log-likelihood for a BARMAX
  model.
}
\details{
The penalty is applied to the intercept (`alpha`), AR (`varphi`),
  and MA (`theta`) coefficients. The regressor (`beta`) and precision
  (`phi`) parameters are not penalized.
}
\keyword{internal}
