% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_model.R
\name{loglik_arx_ridge}
\alias{loglik_arx_ridge}
\title{Penalized Log-Likelihood for a BARX Model}
\usage{
loglik_arx_ridge(y, ar, X, link, alpha, varphi, phi, beta, penalty)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ar}{A numeric vector of positive integers specifying the AR lags.}

\item{X}{A numeric matrix of external regressors.}

\item{link}{A character string specifying the link function.}

\item{alpha}{The numeric intercept term.}

\item{varphi}{A numeric vector of AR parameters.}

\item{phi}{The positive precision parameter (not penalized).}

\item{beta}{A vector of regression coefficients for X (not penalized).}

\item{penalty}{The non-negative L2 penalty parameter (lambda).}
}
\value{
A single numeric value: the penalized log-likelihood.
}
\description{
Computes the L2-penalized (Ridge) log-likelihood for a BARX
  model.
}
\keyword{internal}
