% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_model.R
\name{loglik_ma}
\alias{loglik_ma}
\title{Log-Likelihood for a BMA Model}
\usage{
loglik_ma(y, ma, alpha = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{theta}{A numeric vector of MA parameters. Its length must match the
length of the 'ma' vector.}

\item{phi}{The numeric positive precision parameter.}

\item{link}{A character string specifying the link function ("logit",
"probit", "loglog", or "cloglog").}
}
\value{
A numeric value representing the total conditional log-likelihood
  of the BMA model for the given data and parameters.
}
\description{
This function computes the conditional log-likelihood of a
  Beta Moving Average (BMA) model.
}
\keyword{internal}
