% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_model.R
\name{loglik_max}
\alias{loglik_max}
\title{Log-Likelihood for a BMAX Model}
\usage{
loglik_max(y, ma, X, link, alpha = 0, theta = 0, phi = 0, beta = 0)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{X}{A numeric matrix of external regressors.}

\item{link}{A character string specifying the link function. Valid options
are "logit", "probit", "loglog", or "cloglog".}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{theta}{A numeric vector of MA parameters. Its length must match the
length of the 'ma' vector.}

\item{phi}{The numeric positive precision parameter.}

\item{beta}{A vector of regression coefficients for exogenous variables.}
}
\value{
A numeric value representing the total conditional log-likelihood.
}
\description{
This function computes the conditional log-likelihood of a
  Beta regression model with Moving Average (MA) errors and exogenous
  regressors (a BMAX model).
}
\keyword{internal}
