% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_model.R
\name{loglik_max_ridge}
\alias{loglik_max_ridge}
\title{Penalized Log-Likelihood for a BMAX Model}
\usage{
loglik_max_ridge(y, ma, X, link, alpha, theta, phi, beta, penalty)
}
\arguments{
\item{y}{A time series of numbers in the interval (0,1).}

\item{ma}{A vector specifying the moving average (MA) component.}

\item{X}{A matrix of external regressors.}

\item{link}{The link function ("logit", "probit", "cloglog").}

\item{alpha}{The intercept term of the linear predictor.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the beta distribution.}

\item{beta}{A vector of regression coefficients for external regressors.}

\item{penalty}{A non-negative scalar for the L2 penalty.}
}
\value{
The value of the L2 penalized log-likelihood.
}
\description{
This function computes the L2 penalized log-likelihood of
  the Beta Moving Average Regression (BMAR) model.
}
\keyword{internal}
