% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_model.R
\name{score_vector_ar}
\alias{score_vector_ar}
\title{Score Vector for a BAR Model}
\usage{
score_vector_ar(y, ar, alpha = 0, varphi = 0, phi = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) component.}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}
}
\value{
The score vector of the BAR estimators.
}
\description{
This function computes the score vector of the Beta
  Autoregressive (BAR) model.
}
\keyword{internal}
