% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_model.R
\name{score_vector_ar_ridge}
\alias{score_vector_ar_ridge}
\title{Penalized Score Vector for a BAR Model}
\usage{
score_vector_ar_ridge(y, ar, alpha, varphi, phi, link, penalty)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ar}{A numeric vector of positive integers specifying the AR lags.}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{varphi}{A numeric vector of AR parameters.}

\item{phi}{The numeric positive precision parameter (not penalized).}

\item{link}{A character string specifying the link function.}

\item{penalty}{The numeric regularization parameter (lambda).}
}
\value{
A numeric vector representing the L2-penalized score.
}
\description{
This function computes the L2-penalized (Ridge) score vector
  for a Beta Autoregressive (BAR) model.
}
\keyword{internal}
