% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_model.R
\name{score_vector_arma}
\alias{score_vector_arma}
\title{Score Vector for a BARMA Model}
\usage{
score_vector_arma(y, ar, ma, alpha = 0, varphi = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{A time series of data, with values in the interval (0, 1).}

\item{ar}{A numeric vector of positive integers specifying the AR lags.}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{varphi}{A numeric vector of AR parameters.}

\item{theta}{A numeric vector of MA parameters.}

\item{phi}{The numeric positive precision parameter.}

\item{link}{A character string for the link function, e.g., "logit".}
}
\value{
A numeric vector containing the score values, ordered as
  (alpha, varphi, theta, phi).
}
\description{
This function computes the score vector for a Beta
  Autoregressive Moving Average (BARMA) model. The score vector is the
  gradient of the log-likelihood function with respect to the model
  parameters.
}
\keyword{internal}
