% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_model.R
\name{score_vector_arma_ridge}
\alias{score_vector_arma_ridge}
\title{Penalized Score Vector for a BARMA Model}
\usage{
score_vector_arma_ridge(y, ar, ma, alpha, varphi, theta, phi, link, penalty)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ar}{A numeric vector of positive integers specifying the AR lags.}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{alpha}{The numeric intercept term.}

\item{varphi}{A numeric vector of AR parameters.}

\item{theta}{A numeric vector of MA parameters.}

\item{phi}{The numeric positive precision parameter (not penalized).}

\item{link}{A character string specifying the link function.}

\item{penalty}{The numeric regularization parameter (lambda).}
}
\value{
A numeric vector for the L2-penalized score. The order is:
  (alpha, varphi, theta, phi).
}
\description{
This function computes the L2-penalized score vector for a
  BARMA model.
}
\keyword{internal}
