% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax_model.R
\name{score_vector_armax}
\alias{score_vector_armax}
\title{Score Vector for a BARMAX Model}
\usage{
score_vector_armax(
  y,
  ar,
  ma,
  X,
  link,
  alpha = 0,
  varphi = 0,
  theta = 0,
  phi = 0,
  beta = 0
)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) lags.}

\item{ma}{A vector specifying the moving average (MA) lags.}

\item{X}{A matrix of exogenous variables (regressors).}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A vector of regression coefficients for exogenous variables.}
}
\value{
The score vector of the BARMAX estimators.
}
\description{
This function computes the score vector of the Beta
  Autoregressive Moving Average model with exogenous regressors (BARMAX).
}
\keyword{internal}
