% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armax_model.R
\name{score_vector_armax_ridge}
\alias{score_vector_armax_ridge}
\title{Penalized Score Vector for a BARMAX Model}
\usage{
score_vector_armax_ridge(
  y,
  ar,
  ma,
  X,
  link,
  alpha,
  varphi,
  theta,
  phi,
  beta,
  penalty
)
}
\arguments{
\item{y}{Time series data in (0,1).}

\item{ar}{Vector of AR lags.}

\item{ma}{Vector of MA lags.}

\item{X}{Matrix of external regressors.}

\item{link}{Link function string.}

\item{alpha}{Intercept term.}

\item{varphi}{Vector of AR parameters.}

\item{theta}{Vector of MA parameters.}

\item{phi}{Precision parameter (not penalized).}

\item{beta}{Vector of regression coefficients (not penalized).}

\item{penalty}{L2 penalty parameter (lambda).}
}
\value{
A numeric vector for the L2-penalized score. The order is:
  (alpha, varphi, theta, phi, beta).
}
\description{
Computes the L2-penalized score vector for a BARMAX model.
}
\keyword{internal}
