% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_model.R
\name{score_vector_arx}
\alias{score_vector_arx}
\title{Score Vector for a BARX Model}
\usage{
score_vector_arx(y, ar, X, alpha = 0, varphi = 0, phi = 0, beta = 0, link)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ar}{A vector specifying the autoregressive (AR) component.}

\item{X}{A matrix of exogenous variables (regressors).}

\item{alpha}{The intercept term.}

\item{varphi}{A vector of autoregressive (AR) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A vector of regression coefficients for exogenous variables.}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}
}
\value{
The score vector of the BARX model estimators.
}
\description{
This function computes the score vector of the Beta
  Autoregressive model with exogenous regressors (BARX).
}
\keyword{internal}
