% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_model.R
\name{score_vector_ma}
\alias{score_vector_ma}
\title{Score Vector for a BMA Model}
\usage{
score_vector_ma(y, ma, alpha = 0, theta = 0, phi = 0, link)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{theta}{A numeric vector of MA parameters.}

\item{phi}{The numeric positive precision parameter.}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}
}
\value{
The score vector of the BMA estimators.
}
\description{
This function computes the score vector of the Beta Moving
  Average (BMA) model.
}
\keyword{internal}
