% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_model.R
\name{score_vector_ma_ridge}
\alias{score_vector_ma_ridge}
\title{Penalized Score Vector for a BMA Model}
\usage{
score_vector_ma_ridge(y, ma, alpha, theta, phi, link, penalty)
}
\arguments{
\item{y}{A time series object of data in (0,1).}

\item{ma}{A numeric vector of positive integers specifying the MA lags.}

\item{alpha}{The numeric intercept term in the linear predictor.}

\item{theta}{A numeric vector of MA parameters.}

\item{phi}{The numeric positive precision parameter (not penalized).}

\item{link}{A character string specifying the link function.}

\item{penalty}{The numeric regularization parameter (lambda).}
}
\value{
A numeric vector representing the L2-penalized score.
}
\description{
This function computes the L2-penalized (Ridge) score vector
  for a Beta Moving Average (BMA) model.
}
\keyword{internal}
