% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_model.R
\name{score_vector_max}
\alias{score_vector_max}
\title{Score Vector for a BMAX Model}
\usage{
score_vector_max(y, ma, alpha = 0, theta = 0, phi = 0, beta = 0, X, link)
}
\arguments{
\item{y}{Data, a time series of numbers in (0,1).}

\item{ma}{A vector specifying the moving average (MA) component.}

\item{alpha}{The intercept term.}

\item{theta}{A vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the Beta distribution.}

\item{beta}{A vector of regression coefficients for exogenous variables.}

\item{X}{A matrix of exogenous variables (regressors).}

\item{link}{The link function ("logit", "probit", "loglog", "cloglog").}
}
\value{
The score vector of the BMAX model estimators.
}
\description{
This function computes the score vector of the Beta Moving
  Average model with exogenous regressors (BMAX).
}
\keyword{internal}
